Moved web server into subdir in preparation to also have web client.  Changed necessary paths.
