+++ .gitignore
@@ -1,1 +1,3 @@
-node_modules/*
+node_modules
+npm-debug.log
+
