+++ package.json
@@ -2,7 +2,9 @@
   "name": "balance-transfer",
   "version": "0.0.1",
   "description": "A balance-transfer example node program to demonstrate using NodeSDK apis with alpha based fabric images",
-  "main": "app/app.js",
+  "scripts": {
+    "start": "node web/server/app.js"
+  },
   "keywords": [
     "fabric-client sample app",
     "balance-transfer node sample",
