+++ runApp.sh
@@ -56,4 +56,4 @@ restartNetwork
 
 installNodeModules
 
-PORT=4000 node app
+PORT=4000 npm start
