+++ web/server/app.js
@@ -28,14 +28,14 @@ var expressJWT = require('express-jwt');
 var jwt = require('jsonwebtoken');
 var cors = require('cors');
 var winston = require('winston');
-var config = require('./config.json');
-var helper = require('./app/helper.js');
-var channels = require('./app/create-channel.js');
-var join = require('./app/join-channel.js');
-var install = require('./app/install-chaincode.js');
-var instantiate = require('./app/instantiate-chaincode.js');
-var invoke = require('./app/invoke-transaction.js');
-var query = require('./app/query.js');
+var config = require('../../config.json');
+var helper = require('./helper.js');
+var channels = require('./create-channel.js');
+var join = require('./join-channel.js');
+var install = require('./install-chaincode.js');
+var instantiate = require('./instantiate-chaincode.js');
+var invoke = require('./invoke-transaction.js');
+var query = require('./query.js');
 
 var host = process.env.HOST || config.host;
 var port = process.env.PORT || config.port;
