+++ web/server/helper.js
@@ -27,7 +27,7 @@ var Peer = require('fabric-client/lib/Peer.js');
 var copService = require('fabric-ca-client/lib/FabricCAClientImpl.js');
 //var FabricCAServices = require('fabric-ca-client/lib/FabricCAClientImpl');
 var FabricCAClient = copService.FabricCAClient;
-var config = require('../config.json');
+var config = require('../../config.json');
 var hfc = require('fabric-client');
 hfc.addConfigFile(path.join(__dirname, 'network-config.json'));
 var ORGS = hfc.getConfigSetting('network-config');
