+++ web/server/query.js
@@ -21,7 +21,7 @@ var hfc = require('fabric-client');
 var utils = require('fabric-client/lib/utils.js');
 var Peer = require('fabric-client/lib/Peer.js');
 var EventHub = require('fabric-client/lib/EventHub.js');
-var config = require('../config.json');
+var config = require('../../config.json');
 var helper = require('./helper.js');
 
 var logger = helper.getLogger('Query');
